- Вернул возможность перезагружать голосовой чат - Обозначены территории банд на карте (https://dednet.ru/map) - Обновлены ранги в академии Шериф Департамента - Добавлены переводы E-Coin на другие кошельки - Пополнил список транспорта для покупки гос. организаций - Фикс бага с голосовым чатом, когда на секунду бывает слышно других людей - Фикс бага с тюрьмой и смертью персонажа - Фикс прокачки навыка пилота - Правки по админке

+++ src_client_package/src/jobs/bus.js
@@ -183,6 +183,12 @@ bus.start = function(busType) {
 };
 
 bus.nextCheckpoint = function() {
+
+    if (mp.players.local.vehicle.getPedInSeat(-1) !== mp.players.local.handle) {
+        mp.game.ui.notifications.show('~r~Необходимо быть на водительском месте');
+        return;
+    }
+
     jobPoint.delete();
 
     let timeout = 1;
